Fix browser address
It now properly displays as '/' for the main page. To do this I had to get rid of Authorization container and move those contents into DisplaySaves. Now the main page will redirect the user if they're not signed in. Also, Initialization container renamed to Pagination because the initial loading logic was moved to a new container called SignIn.